Show mex how i invisaged SSL negotioation.
Both the server and the client have 3 modes of SSL "preference":
a) always, b) if possible, c) never.

netd and the client will have a conversation and decide if they are compatible.

States:
server a + client a = OK + SSL.
server a + client b = OK + SSL.
server a + client c = ERROR.
server b + client a = OK + SSL.
server b + client b = OK + SSL.
server b + client c = OK + NOSSL.
server c + client a = ERROR.
server c + client b = OK + NOSSL.
server c + client c = OK + NOSSL.

In any ERROR case we can't continue and exit.

In *any* case with NOSSL, a warning is spewed onto the console. No SSL is a bad
thing, but is not strictly an error.

Once these conditions are checked, and the ssl connection is opened (if needed),
a simple check for ssl == NULL is all that is needed to know whether we are
encrypting traffic or not.

Used #defines instead of enum purely for consistency, no real preference is
reality.